requirements.txt
# Web / API
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0

# CLI
typer[all]==0.12.3
rich==13.7.1

# Storage (boto3 handles both R2 and MinIO via S3-compatible API)
boto3==1.34.101
botocore==1.34.101

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
pgvector==0.3.2
psycopg2-binary==2.9.9
alembic==1.13.1

# AI / Embeddings
sentence-transformers==2.7.0
torch==2.3.0
transformers==4.41.1

# OCR / File Parsing
pytesseract==0.3.10
Pillow==10.3.0
PyMuPDF==1.24.3
python-docx==1.1.2
openpyxl==3.1.2

# File Watching
watchdog==4.0.1

# Config
python-dotenv==1.0.1

# Async utils
aiofiles==23.2.1

# Note: torch version may conflict with older sentence-transformers; pin both for consistency
